services/demucs/requirements.txt
torch
torchaudio
demucs
mutagen
pydub
python-dotenv

torch==2.7.1
torchaudio==2.7.1
demucs==4.0.1
mutagen==1.47.0
pydub==0.25.1
python-dotenv==1.1.1
